refactor: move components and views
